Updates Labels to be able to include "arguments"
Expanded and renamed the "timestamped" label classes to also
potentially include "arguments" -- values set within a label that
prescribe something about what is being labeled.  This will be
essential for continuously parameterized gates, which we plan to
implement by having layer lizard objects construct argument-containing
labels using new operator-factory objects (future work).  This
commit includes updates to the circuit parser so that you can read
in circuits with timestamps and arguments using the "!" and ";"
syntax, respectively.  For example: "GrotX;a4.5:0:1!0.2" describes
a gate with a single parameter: "a4.5", sslbls=(0,1), and a
timestamp of 0.2.