Adds framework for "operation factories" to support gates w/args.
This commit adds initial support for continuously parameterized
gates, or "gate with args", corresponding to Label objects with
arguments.  The idea is that Model objects (so far just
 LocalNoiseModel) will contain "operation factory" objects (see
opfactory.py) which can be called upon by a LayerLizard to
construct an operation based on either or both of a list of state-
space labels and a list of arguments (coming from the originating
label).

Within opfactory.py there is the OpFactory base class,
which is meant to be derived by user classes, as well as two
general-purpose factory classes: EmbeddedOpFactory and
EmbeddingOpFactory.  The former allows a factory that constructs
operation on some space (usually 1 or 2 qubits) to be cast as a
factory on the entire state space by simply embedding the
operations in some predetermined way.  The EmbeddingOpFactory
constructs on-demand embeddings based on target labels requested
from the originating label - i.e. it's create_op method expects
a non-None `sslbls` argument to direct it as to how the underlying
operation should be embedded.  This would allow, for example, a
gate to be available anywhere in system of qubits without needing
to construct all possible embeddings ahead of time.

Very little testing has been done to this point.  This commit also
includes a bug fix in Label.__new__(...).