Adds ability to easily initialize a Circuit from a string.
Now, if the intial argument of Circuit.__init__ is a string, it
is parsed as a circuit and the resulting circuit constructed.
Previously one would need to set this initial argument to None and
set `stringrep` (which is still an option) to acheive this
behavior -- this new way is much simpler.