Make `time` a non-value-changing property of all Labels.
This commit revamps label.py so that instead of having special
Label classes for whenever time *or* arguments are specified,
*all* label objects hold a time and the additional classes are
only for when arguments are needed.  This change is primarily a
result of having a better understanding of what role the 'time'
(previously 'timestamp' in some places) of a Label represents.

It seems that this value should represent the duration of the label,
but that this duration shouldn't change how the Label is tested
for equality or hashes.  That is, a Label is identified by its
name, sslbls, and (optionally) arguments -- it can be tagged with
a time/duration as meta-data, but this is not used to distinguish
it from other labels.  This fits with the fact that a Label's time
is used when computing circuit probabilities, to compute the time
at which operations occur but not what those operations are.

There is thus a parallel between the fact that a *single* LinearOperator
or SPAMVec object is used at whatever time it occurs (this
time is set via its set_time method) and a single Label can have
any time value and still be the *same* label:  Labels correspond
to operators / spamvecs.

To implement this, this commit removes the time value from the
underlying tuple that represents most Labels (and is used in
equality testing and hashing) and make this a separate property.
This naturally leads to all Labels having this time property and
only needing separate classes for Labels with arguments, which
correspond to the cases when the underlying tuple needs to change.