Moves time-dependent computation code to replibs with slight refactor.
In preparation for creating faster C-versions of some of the DM_*
time-dependent computation functions, this commit moves this
functions from mapforwardsim.py into slowreplib.py and fastreplib.pyx.
Right now, the version in fastreplib.pyx is identical to the slow
version, so there's not much speedup but the code still runs and
checks out in the time-independent case.

The `DM_compute_TDchi2_cache` and `DM_compute_TDloglpp_cache`
functions have been refactored (in both replibs) to use a common
(and new) `DM_compute_TDcache` function, which eliminates many
lines of repeated code.

Finally, the `spamTupleToOutcome` function has been moved from
model.py to optools.py as this is used by both Model objects
and replib functions - so this seems like a better place for it.