Fixes minor bugs; integrates UnitaryOpFactory; adds ComposedOpFactory
In addition to fixing minor bugs introduced in recent commits,
this commit integrates UnitaryOpFactory by enabling the user to
specify a function as an element of `nonstd_gate_unitaries` that
takes a tuple of arguments and returns a unitary numpy array
(matrix).  An important caveat that isn't tested for well now
is that this function must also respond to being passed `None`
by returning some default unitary so that construction routines
can determine the dimension of the unitary matrices it produces.

This commit also adds the ComposedOpFactory class which as its
name implies composes zero or more factories or operators.
The build_crosstalk_free_model function utilizes this factory
to add an error map based on supplied error rates to the output
of a user supplied factory (possibly through nonstd_gate_unitaries).