Makes a few tweaks to increase drift code speed.
Adds a new `DataSet.merge_outcomes` function which I think is
correct (we should compare with the similarly named function in
datasetconstruction.py) and is a bit faster.  Also minor tweaks
in the drift code and DataSet methods to try to eek out a little
extra speed based on profiling results.