Fixes bugs causing a drift test to fail under python2.7.
Two bugs fixed:
1) the tuple keys of a stability analyzer's _tupletoindex member
rely on a specific ordering that is not guaranteed when running
tuple(dict.values()) within python2.7 (this works in python3).
2) Calling class methods without an instance is not allowed in
python2.7, and so calls to ProbTrajectoriesPlot and PowerSpectraPlot's
_create functions from within their Germ-Fiducial variants now
create instances of these classes first.  It seems like the Germ-Fid
variants may actually want to inherit from ProbTrajectoriesPlot and
PowerSpectraPlot in the future (?).