Initial commit on improving pruned-path calcs: adds max_npaths  and required_achieve_factor.
The content of this commit may largely be reverted as it doesn't seem
to work very well - but we're committing it here so we have record
of it.  This commit adds two main functionalities to the pruned-path
calculations:

1) adds a max_npaths argument to a TermForwardSimulator object,
that gets propagated down into slowreplib.py and caps the number
of paths that will be traversed per circuit outcome.  This seemed
useful because, when optimizing, initial "wild" steps can explore
large-error regimes where the thresholding alone would lead to tons
of paths being traversed - more than we'd want.

2) we've noticed a second issue that can occur when optimizing with
pruned-paths: the optimizer will continue to make steps that only
improve the objective because the objective is being truncated by,
e.g. max_npaths or a min_threshold.  The result is that the gap between
the achieved and target sum-of-path-magnitudes keeps growing, while
the objective function is modestly decreased.  We want to avoid this
regime but aren't sure how yet -- adding a 'required_achieve_factor'
attempts to flag this condition by asserting failure when some
potentially large factor times the achieved SOPM is still less than
the target SOPM.

Finally, we note that all "replib" parts of this have only been
implemented in the slow Python version (in slowreplib.py) and haven't
been transfered to fastreplib.pyx yet (because things are still in
flux).