Initial commit on task of integrating representations & operations.
The big idea is that we want to eliminate all .torep() calls in
favor of having representation objects stored internally within each
operation object such that there is a 1-1 correspondence between these
objects at all times (when the operation is updated, so is its
representation object - the *same* object is updated in a persistent
way, a new rep object never needs to be constructed).

This commit attempts to make this change for all the operation
classes in operation.py.  No testing has been done so there's likely
a lot of dumb errors.  The next steps are to do the same think for
the SPAM ops in spamvec.py and then remove/alter all the .torep calls
that occur in the forwardsim objects.

Another issue that we'll need to deal with later is pickling the
rep objects (we've avoided this so far by having them be ephemeral).