Updates .torep() -> ._rep in map forward sim.  Fixes minor bugs.
In particular, updates DenseOperator and DenseSPAMVec so that
ProtectedArray objects are handled specially and the self.base
member is set to a ProtectedArray rather than the raw contiguous
array version of the argument given to __init__.

This commit also updates the self.base member of DenseSPAMVecs so
instead of being a (N,1)-shaped 2D array it's a (N,)-shaped 1D
array to exactly reflect that stored in self._rep.  This meant
updating lots of self.base[:,0] to self.base[:], etc., in spamvec.py.

Testing has begun, bug got hung up on needing __reduce__ methods
for the rep classes just to *copy* them (not just to pickle them), so
we'll need to get that up and running next.