Reverts recent change to DenseSPAMVec shape & fixes deep copying.
This commit returns the .base member of a DenseSPAMVec to being
a (N,1)-shape array even though the _rep holds a 1D array.  This
makes deep-copying more difficult -- in particular the TPSPAMVec,
when using the default deep copying, doesn't identify the array
held by the _rep with the ProtectedArray held as self.base.  This
required a hack/custom __deepcopy__ method for TPSPAMVec that fixes
these memory pointers.  I'd like to find a better way of rigging
this in the future -- maybe moving to having spam vecs that are
just 1D arrays like the reps?