Fixes bugs to get GST working again; begins optimizations.
This commit completes the initial implementation of having
"integrated reps", so that now all the different GST modes run
again and seem to be working correctly (i.e. they give reasonable
results and agree with one another).  The routines don't run much
faster and may run slower than when we started now - there will
need to be some optimization performed next.  There's also a number
of debugging artifacts that still remain, so we'll need to polish
this up too.  Finally, we'll need to update slowreplib.py to
match the current (or future) fastreplib.pyx.

Here are the main bug fixes/improvements/experimental ideas of
this commit:
- declares a bunch of Cython class members as "public" and names
them more intelligently (not all just "data_refX") so debugging
reps is easier.
- adds needed set_exp_params method of DMOpRep_Lindblad
- term reps have been updated: there's still a RankOneTerm.torep(...)
 function, but it only needs a single "max-number-of-poly-vars"
 argument; the RankOneTerm contains it's own termtype (prep, effect,
 or gate) and now has an _evotype member so there's no more "dense"
 and "clifford" term types.  Calculation routines still call torep
 for terms, but in future enhancements should just cache
 the reps and not the term objects themselves.
- comments out RankOneTerm.collapse method which I think is never
 used.
- because Model._clean_paramvec seems to take a long time in profiling
 results, this commit adds a "nodirty" argument to many/all of the
 various from_vector methods so that we can signal these methods to
 *not* set self.dirty=True when from_vector is called in a known
 context, which may be the cause of unnecessary _clean_paramvec calls.
 This could also be barking up the wrong tree and need to be removed
 later.