Bug fixes and continued optimization of map-based calculations.
Comparison with last stable version revealed a bug caused by
setting "nodirty=True" in MapForwardSim's from_vector method.  This
I think is due to the fact that we need dirty flags to be set when
a fwdsim calls "from_vector" since this *doesn't* update the parent
Model's paramvec and the Model needs to know to sync this later on.

This commit leaves the "nodirty" flags present in many/all of the
from_vector functions, but it isn't used anywhere (always False)
and should probably be removed unless we find another use for these
flags soon.

Alongside the "nodirty" flags, this commit adds "close" flags to
the from_vector functions so that the caller can indicate if the
vector is *very close* (e.g. in finite-diff derivatives) to the
current vector.  This allows LindbladOp objects to bypass recomputing
the m_star and s quantities needed for sparse matrix exponentiation
that are integers and based only on matrix norms which will not have
changed appreciably.  This does result in a speedup for 2Q-GST tests,
where the m_star and s computation actually dominates the computation
time - especially when errcomp_type='gates' is used.

At this point, more optimization is still needed but the results
from all the different reduced-model 2Q-GST modes we test (in calc
demo notebook) agree with those of the last stable version (0.9.8),
so it looks like the code is correct now.