Adds _iter_matrix_elements optimization to EmbeddedOp.
For map and matrix calcs using dense operations the todense operation
of EmbeddedOp was a bottleneck due to the loops in _iter_matrix_elements(...).
This commit organizes the _iter_matrix_elements code a little more
and adds a `_iter_elements_cache` member to EmbeddedOp that simply
caches the values to iterate over.  This greatly speeds subsequent calls
to _iter_matrix_elements, but may cause memory issues for higher qubit
counts so we'll need to beware of this in the future.  Note that for
higher qubit counts we'd probably want to use *sparse* objects which
should avoid this issue altogether.