Adds integrated-term and "nodirty" optimizations to map calcs.
To speed up map-based calcs this commit moves some operation-building
(LayerLizard.get_operation calls, etc) and what was previously
.torep calls (now ._rep references) outside of the finite-differencing
loop within DM_compute_dpr_cache in fastreplib.pyx.  This type of
optimization is what this whole endeavor of integrating representation
objects with their container operations was for - it's nice to see
some benefits.  Not well tested yet: if there are issues it's likely
because in the finite difference loop the .from_vector calls (which
ultimately get to LayerLizzard.from_vector) may not call from_vector
on all the relevant operations - just the "primitive" operations of
the model as known by the layer lizard.  The fix for this would
be to have LayerLizard.from_vector be given (or know about) the list
of operations being stored in DM_compute_dpr_cache, etc.

The second optimization is also within DM_compute_dpr_cache, and
is to set nodirty=True for the from_vector calls within the inner
finite differencing loop.  This will prevent _clean_paramvec from
being called unnecessarily amidst the calculations, e.g. if there's
a reference to model.num_params.