Adds sparse matrix sum optimization.
Adds new csr_sum_flat function to matrixtools.py which supercedes
the csr_sum function but further streamlining the computation at
hand.  The resulting function is able to be succinctly implemented
in Cython (in fastcalc.pyx).  The old code is commented out and
tagged with REMOVE.  The resulting tests show the for map-based
calcs using sparse objects the vast majority of the CPU time is
spent on forward simulation like it should be.  The biggest bottleneck
after this is the LindbladErrorGen._update_rep function, which
performs the now-faster csr summing.