Adds 'acton' methods to LinearOperator, POVM, and Instrument.
These single-operation functions allow easy access to performing
serial operations on 'prep'-type SPAMVecs, which are essentially
states.  This allows pyGSTi to be used in a mode similar to other
forward-simulation toolkits like quantumsim or quest.  Instruments
can only be used with the "densitymx" evotype so far, and this is
generally still a work in progress.