Refactors do_mlgst and do_mc2gst routines; updates term-poly caching.
This commit continues on the thread of improving term-based
calculations and incorporating path pruning in a meaningful way.
It has been determined that it's not a good idea to prune paths in
the middle of an optimization loop - better to choose a set of paths
(pruned or just a fixed Taylor order, for instance) and optimize,
then repeat.  This intended functionality prompted two separate
but related updates found in this commit:

1) since we may want to refine this path pruning over
multiple iterations with the same set of circuits (i.e. the same
maxL value), this commit factors out the innermost optimization
call in do_mc2gst and do_mlgst into the new _do_runopt(...) function.
This also consolidates repeated code between the two functions.  As
of now, _do_runopt is called only once, but in near-future commits
it can be called in a loop within do_mlgst or do_mc2gst.

2) since we don't want to change the path truncation amidst an
optimization loop, this commit introduces the forward-simulator
bulk_prep_probs function which is called to compute things needed
to call bulk_fill_probs, etc.  This means the probability-polynomials
for term-based calculations, and so termforwardsim.py and
termevaltree.py have been updated accordingly.  Now there's a
all_p_polys dictionary in TermEvalTree that get's populated (once)
by bulk_prep_probs before an optimization and computes (if needed)
the polynomials required given the current model (which could impact
how much to prune, etc.).  Then, in calls like bulk_fill_probs, just
the all_p_polys cache is accessed - NO attempt is ever made to compute
new polynomials.  This is all a bit hacky now, which multiple
caches in the TermEvalTree, but the basic framework is there; we'll
need to neaten this up in the future.