Removed GPL Trove tag from setup.py, added Apache License.
Note that the approved Trove tag for the Apache License 2.0 is
    License :: OSI Approved :: Apache Software License
Which is a little abiguous/confusing given that the actual title of the license
is "the Apache License, Version 2.0". In principle this classifier could also
refer to earlier versions of the Apache License (1.0 and 1.1); however, only
version 2.0 is OSI approved, so really it's just a mess.

Relevant discussion: https://github.com/pypa/pypi-legacy/issues/91