Fixes bugs with owned-memory in dense reps (unsettable writable flag).
Because we didn't ensure that the memory being held by DenseSPAMVec
and DenseOperator was *owned* by that array (in numpy's sense of
ownership), the ndarray's writable flag wasn't always able to be
set to True after it had been set to False.  This commit adds
numpy.require calls to ensure that the memory held by these objects
(and shared by their contained representation objects) is owned
and so their .base.flags.writable values can be modified at will.

This commit also fixes a small bug in that gate-based GaugeGroup
objects required a LinearOperator and were sometimes passed a plain
numpy array (by LGST).  This commit adds logic that converts numpy
arrays to StaticDenseOps within GateGaugeGroup so that this works
again (this was broken when we added .todense() calls recently,
which assume we have a LinearOperator and not just a numpy array).