Initial work of updating map-based forward sim & eval tree.
The way in which the map-based forward simulator computes
probabilities can be inefficient, particularly for sparse data.
To conform to the matrix-fwdsim type of evaluation, where computing
probabilities for multiple (rho,E) pairs is easy once you've computed
the full superoperator, the map-based (and term-based) forward
simulators computed all their circuit outcomes for the entire set
of E-labels that were present in the evaluation tree (often = the
dataset).  This is fine when all the circuits have counts for the
same set of outcome labels, but for multi-qubit situations the
data can be quite sparse and this means of computing is wasteful.

This commit begins a restructuring of the way the map-based calculator
performs calculations, evaluating only the outcomes labels that
are required.  Still a WIP, and so far this work has been done by
adding members to the MapForwardSim object - members that will be
moved to the representation libs when they're working better.  Also,
this commit leaves OLD_* versions of a number of fwdsim functions
in for reference; these will obviously be removed soon.