Updates (yet again) DenseSPAMVec to hold a (N,)-shaped array.
Because holding a ._rep and .base that are different shapes and
a .base that is a non-native numpy array (e.g. a ProtectedArray)
is so difficult, this commit initializes a DenseSPAMVec with a
.base1D member that is a normal contiguous numpy array.  The .base
member is now a property that returns a (N,1)-shaped view of
.base1D, which makes this update compatible with (I think) everything
that depends on .base being a (N,1)-shaped array.  In the case of
TPSPAMVec the .base property is overloaded to return an appropriate
ProtectedArray.

This commit also updates .base references in spamvec.py to .base1D
references wherever possible.