Organizes new mapforwardsim.py routines; ready for move to replib.
This commit organizes the new calculation routines into methods
such as _fill_probs_block(...) and gets these routines ready for
migration into the replibs for performance.  These routines for
computing probabilities still pass around a lot of index arrays and
often convert slices to index arrays - maybe in the future we could
go through and try to allow slices to be processed more efficiently.

This commit also fixes an integer test in matrixtools.py to include
the np.int64 type, which isinstance(., int) returns False for.