Updates circuit-simplification paradigm and moves map-calc routines.
This work started out attempting to move core computation routines
from MapForwardSimulator to the replibs, but mid-way through it
became apparent that the way in which we "simplify" (previously
"compile") circuits was tailored to matrix based calculations.
Namely, holding a "raw_dict" of spam-tuples, i.e. (rhoLbl, eLbl)
pairs is convenient when an operation's process matrix is computed
using products, as performing the SPAM contraction is the last step.
For map-based calculations, however, the selection of rho is the
*first* step, and so the prep label should be included with the
circuit rather than held separately.  As such, this commit alters
the "simplified-circuit" representation so that we transform a
general (w/POVMs and instruments, etc) circuit into a "raw_dict"
whose keys are simple circuits that include prep layers and whose
values are lists of just effect-labels instead of spam-tuples.

MatrixEvalTree has been updated to add a processing step
that strips prep-layers off and still holds a simplified_circuit_spamTuples
member, wheras MapEvalTree now holds a simplified_circuit_elabels
member instead.  (This means that the forward simulators other than
MatrixForwardSimulator can't use the _fill_result_tuple_* functions,
but these are being phased out now anyway, as they are tailored to
the matrix-based simulator.)

This commit also makes the changes it set out to make: core
computation routines in MapForwardSimulator have been moved to
the replibs (slowreplib.py and fastreplib.pyx), and have been
tested to reproduce a standard map-based 1Q-GST result.

As a side-effect of needing to add a prep-layer to an empty string
when simplifying circuits, Circuit.insert_layer has been updated
so that it adds .line_labels when .line_labels is initially
empty (which is the case for an empty circuit, and should be since
we don't know what will be added to it - e.g. we don't want line_labels
to be ('*',) for instance).