Adds fastreplib.pyx versions of time-dependent obj. function routines.
These are used by the MapForwardSimulator when evaluating the logl
or chi2 objective functions in a time-dependent manner (i.e. with
a potentially time dependent model and data).  The fastrep versions
have been checked to give the same results as the slowrep version,
which gives the same results as what we had before we "upgraded"
the functions to use map-based tailored algorithms.

It's also worth noting that the fastrep versions don't differ so
much from the slowrep versions, and as a result there's just a
nominal 30% speedup by using the fastrep library in the 1Q
case we've been testing.

This commit also makes trivial updates to slowreplib.py just to
maintain consistent function naming with fastreplib.pyx