Upgrades TermForwardSimulator and TermEvalTree; other small fixes.
This commit updates the term-based forward simulator and related
evaluation tree.  Similar to the map-based case, this update
removes references to the matrix-tailored _fill_result_tuple_*
functions resulting in a cleaner and more efficient evaluation.
This update also changes the "evaluation order" of the term-based
tree: it used to sort the circuits lexographically for no good
reason (it doesn't, and I don't think it even can use prefixing),
and now it doesn't -- but leaving the circuits in their original
order, the "final element indices" of a given circuit are particularly
easy to compute, and this makes creating arrays of per-element
quantities easy and efficient.  The upgraded version also uses
simpler caching, and in particular doesn't need the (now commented-
out) functions get_dp_polys and get_hp_polys.

There have been other small bug fixes to code in the map-based
case (discovered by writing the term-based analogue).

The update didn't require any reworking of the replib functions,
and testing a 1Q example showed the same results as we had gotten
prior to beginning the update, which is a good sign.