Removes OLD and unecessary routines from fwdsims and replibs.
Now that fwdsim 'upgrade' is complete, this commit cleans us a
number of the biggest abandoned/irrelevant code segments.  It moves
_fill_result_tuple into MatrixForwardSimulator, as it's only used there,
and removes OLD MapForwardSimulator routines (kept around just for reference).
It also removes the _compute_XXX_cache methods of MapForwardSimulator,
and the DM_compute_*_cache routines in slowreplib.py and fastreplib.pyx.