Adds rough framework for a pruned-path method that might actually work.
This commit builds upon the features of the last one, and assembles
a heuristic algorithm within do_mc2gst and do_mlgst that iteratively
adds paths until the converged output of a leastsq optimization
also meets the pathmagnitude-gap requirement set by the user.  The
basic idea is that we iteratively try to optimize the objective
function using some number of paths, adding paths and trying again
when the output of an iteration has errors too large for the current
number of paths to capture adequately (measured by the given
pathmag-gap). There's still much cleanup and optimization to be
done, as well as figuring out a way to handoff to the next outer (L)
iteration.