Optimizes pruned-path calcs and removes some debug messages.
This commit takes the basic framework (for automatically pruning
path integrals) of the last commit and 1) cleans it up some and
2) optimizes it some.  The Model object now has a
bulk_probs_num_term_failures method with a boolean 'adaptive'
argument that takes the place of calling bulk_prep_probs with weird
flags or of calling internal methods of TermEvalTree to get the
current number of "failures", i.e. circuits whose probabilities are
not adequately computed by the current pruned-path-polynomials
(when adaptive=False) or the best pruned-path-polynomials we can
find using the gap, max-paths, etc. constraints given to the
TermFormwardSimulator.

This commit also adds what was lacking in fastreplib.pyx (namely,
the SV_circuit_pathmagnitude_gap function) so that the Cython
routines can again be used for the entire automated process of
iteratively finding a good set of paths to keep.  Additionally,
the pruned-path-poly computation routines, e.g.
SV_prs_as_pruned_polys, now have a boolean "compute_polyreps"
argument which can be set to False when we're only wanting to
assess how well a given threshold would perform (i.e. the
number of failures) rather than actually needing to compute
the polynomials.  This allows these routines to short-circuit
some unecessary computation for increased performance.

Some refactoring, e.g. the creation of the
_interpolate_model_until_no_failures helper function in core.py,
has been done to clean up the code, and a number of the debug
print statements have been commented out or removed.

Finally, this commit updates the custom-LM solver in a mostly
costmetic way: now linear solver and "no man's land" failures
(still triggered by just a ValueError!) are printed parenthetically
within the "Rejected" message rather than on their own line.  This
cleans up the output a little without sacrificing any detail in
the messaging.