Adds more optimization of pruned path integral calculations.
This commit updates the Polynomial class so that it is very closely
tied to its PolyRep "representation".  The PolyRep class has been
updated accordingly in fastreplib.pyx but NOT in slowreplib.py yet
(so this will need to be done soon).  The new implementation has been
tested against the old implementation, which has been temporarily
left as the SLOWPolynomial class within polynomial.py.

Among updates to the Polynomial class is the addition of
"mapvec_indices[_inplace]" functions to the previously
existing "map_indices[_inplace]" functions.  Index mapping was found
via profiling to be a bottleneck and the new functions reduce this
substantially by being supplied with a vector/array mapping instead
of a Python function.

This commit also makes the `compose_terms` function faster by
creating just a single RankOneTerm (by calling the
 RankOneTerm.__new__ method) instead of constructing many RankOneTerms
the usual way.  The RankOneTerm class is also given a new
mapvec_indices_inplace function (which just delegates to the polynomial
coefficient's function of the same name).

These changes were all initiated by profiling 1- and 2-qubit term-based
GST.  This continues to be a WIP.