Adds more pruned path integral optimization.
We experimented with different ways of using the "restrict_to"
argument introduced in the last commit, and found that this didn't
actually improve anything (and should be taken out later).  Alternate
_interpolate_model_until_no_failures(...) (included but left commented
out in the core.py file - to remove later) functions where tested
and ultimately a variant of the original version, which doesn't use
any "restrict_to" functionality, seemed best (when tested on up to
4 qubits).  The reason seemed to be that it doesn't cost that much
to test shorter and/or "easier" circuits for whether they meet the
desired pathmagnitude gap, and so trying to hone and only test the
hardest circuits actually cost more time.  The difficulty centers
around the fact that we don't want to compute a SOPM for hard circuits
when there are easier circuits that fail to meet the gap.

An enhancement (still a HACK) in _interpolate_model_until_no_failures
is to not always start with alpha=1.0 but to base the initial alpha
on how large the gap violation is.  Perhaps other criteria could be
used as well for this.  Perhaps scaling back the model until the
gap violation is only a factor of 2 or 10 or something before checking
for viable paths?

Another enhancement is to avoid calling bulk_prep_probs and _do_runopt
when alpha=0 is selected by _interpolate_model_until_no_failures,
since in this case we haven't moved at all.  There is a HACK in here
as well having to do with the fragile cache management that we'll
be fixing in the future (see the _do_term_runopt function).

Some initial cache management was done within termevaltree.py,
where we no longer set self.percircuit_p_polys when just_get_nfailures
is True.  The self.percircuit_p_polys dict should only be set by
a legit bulk_prep_probs call which sets polynomials as well, as
this cache is used by calls to check for no-mans-land within an
optimization loop.