Fixes and begins to clarify caching in pruned-path term calcs.
Adds explanatory comments, and fixes some minor issues and confusions
with the way opcache and repcache were being used in the term calcs.
At the center of this was the realization that the repcache used by
the TermEvalTree is more specifically a cache of the high-magnitude
terms for the current set of operations, and that these *need* to
be cached in order to "lock in" a set of paths to be used.  As such,
the repcache, named "highmag_termrep_cache", should be (and now is)
a member of TermEvalTree since the eval tree is the object that "locks in"
a set of paths to use in subsequent evaluations.  The "opcache"
remains owned by the Model object, as this is a more general purpose
cache containing operations that were created by a layerlizzard object
and might be time-consuming to create.  If the layer lizzard were
persistent, it would make the most sense to have it own the "opcache".

The updates in this commit fix an issue whereby the calls within
_interpolate_model_until_no_failures would *update* the Model's or
TermEvalTree's repcache thereby inadvertently updating the selected
set of paths.  Now TermEvalTree.cache_p_pruned_polys is careful not
to update the tree's percircuit_p_polys or highmag_termrep_cache
members when just_get_nfailures == True (how the routine is always
called from _interpolate_model...).  In the future we should refactor
these routines to make more explicit when paths are being updated
vs. something is being computed using the locked in paths.

Still very much a WIP.