Updates function names and performs mild refactoring for clarity.
Updates the names of functions/methods used to compute pruned-path
calculations, and adds comments, to make it clear when the "locked in"
paths (as set by a TermEvalTree) are being used vs. when the paths
are allowed to be "adapted" (i.e. changed within the fwdsim's
user-supplied contraints like the max paths per circuit).  The
OpModel.bulk_probs_get_num_termgap_failures function, for instance,
does not update the "locked in" or "current" paths held by a
TermEvalTree - only the bulk_prep_probs function does this (and
always computes the polynomials resulting from the path choice too).

The speed and functionality have not substantially changed with this
commit - only the clarity of the code is meant to be improved.