Updates rank-one-term classes in term.py, mainly for clarity.
While this started as an effort to optimize calls to RankOneTerm.torep()
this commit ended up refactoring the class(es) in term.py mainly for
increased code clarity.  In particular, we separate the different
type of RankOneTerms (prep, effect, and op), the different coefficient
types (polynomial or number), and on whether the term keeps track of
its magnitude into separate classes (using a somewhat clever
inheritance hierarchy).  This expands the amount of code but makes for
many fewer if/else switches and makes profiling easier, even if it
doesn't actually speed up the code much.

The bottleneck lies in the __init__ and compose methods of the
RankOnePolyOpTermWithMagnitude class, and to circumvent this would
require pushing these classes to Cython to avoid the python object
creation overhead.  This doesn't seem worth the effort at the moment
- the main impediment is that if we only store a RankOneTerm as it's
"rep" then we need to be able to embed and compose just operation
representations, which isn't well supported.