Adds some minor MPI-related updates to get pruned-paths working.
With these updates, the pruned-path calculations seem to be working
correctly with mulitple CPUs, i.e. they give the same results as
a 1-processor run.  The scaling with #procs is still pretty sub-
linear though, perhaps because of poor tree splitting - we'll need
to dig into this further in the future.  This commit adds some
initial debugging statements.