Restructures term.py classes to be representation-wrappers.
Recently we refactored term.py to be clearer.  This commit restructures
the innards of the RankOneTerm-derived classes so that they are
thin wrappers around a corresponding representation object.  Doing
this allows higher performance and, for instance, the representations
themselves to be used instead of RankOneTerm-derived objects within
the exp_terms_above_mag function.

The term-rep classes need to have a composed(...) classmethod and a
few other simple methods to work with their parent objects - these
updates have been done in fastreplib.pyx and slowreplib.py and a 2Q
test shows they give the same results.

This commit also adds an important optimization to the exp_terms_above_mag
function which avoids calling composed(...) when it is unnecessary.