Updates a compact-poly unit test.
The expected outcome for a compact poly has changed not that we
use the rep-level compact_[complex,real] functions in fastreplib.pyx
or slowreplib.py because the *ordering* of terms when "compacting"
a polynomial has changed.  Previously, the variable-index-tuples
where sorted, so for the poly 2x1^2x2 +  3x1x2 the tuples (1,1,2)
and (1,2) would be sorted ascending as (1,1,2) followed by (1,2).
Now the integer-equivalent (a PolyVarIndex compound integer object)
of these tuples is sorted, and the integer corresponding to (1,1,2)
is higher than that of (1,2) so the ordering in the example would
be reversed: (1,2) followed by (1,1,2).  This required updates to
the asserted correctness in the test_polynomial.py unit test module.